refactor: reorganize module structure with domain/ and storage/
- Create domain/ module for core types:
  - domain/file.rs: File, FileId, RequestTemplate, etc.
  - domain/batch.rs: Batch, BatchId, BatchStatus, etc.
  - domain/request/: Request typestate (state.rs, transitions.rs)
  - domain/daemon/: Daemon typestate (state.rs, transitions.rs)

- Rename manager/ to storage/ for clearer naming
  - storage/mod.rs: Storage, DaemonStorage, DaemonExecutor traits
  - storage/postgres.rs: PostgreSQL implementation
  - storage/utils.rs: Size calculation utilities

- Maintain backwards compatibility:
  - batch/mod.rs re-exports from domain/
  - request/mod.rs re-exports from domain/request/
  - manager/mod.rs re-exports from storage/
