+++ src/lib.rs
@@ -8,19 +8,21 @@
 
 pub mod batch;
 pub mod daemon;
+pub mod domain;
 pub mod error;
 pub mod http;
 pub mod manager;
 pub mod request;
+pub mod storage;
 
 // Re-export commonly used types
 pub use batch::*;
 pub use daemon::{Daemon, DaemonConfig, ModelEscalationConfig, SlaThreshold};
 pub use error::{FusilladeError, Result};
 pub use http::{HttpClient, HttpResponse, MockHttpClient, ReqwestHttpClient};
-pub use manager::postgres::{PoolProvider, PostgresRequestManager, TestDbPools};
-pub use manager::{DaemonExecutor, Storage};
 pub use request::*;
+pub use storage::postgres::{PoolProvider, PostgresRequestManager, TestDbPools};
+pub use storage::{DaemonExecutor, DaemonStorage, Storage};
 
 /// Get the fusillade database migrator
 ///
